+++ bundlers/BrowserifyHelper.ts
@@ -21,11 +21,45 @@ module BrowserifyHelper {
     }
 
 
+    /** Listeners for bundle build/compilation events
+     */
+    export interface BuildListeners {
+        /** When a bundle error of any kind occurs
+         * @param srcName the name of the stream/compilation step where the error occurred
+         * @param dstFileName the file name of the bundle the error occurred on
+         * @param err the error
+         */
+        error?: (srcName: string, dstFileName: string, err: any) => void;
+        /** When all bundles finish building
+         * @param res an object with build results about all the bundles built
+         */
+        finishAll?: (res: BuildResults) => void;
+        /** When a single bundle finishes building
+         * @param fileName the file name of the bundle that finished building
+         */
+        finishBundle?: (fileName: string) => void;
+        /** When a bundle skips being built
+         * @param fileName the file name of the bundle that skipped being built
+         */
+        skipBundle?: (fileName: string) => void;
+        /** When a bundle starts being built
+         * @param fileName the file name of the bundle that is about to begin building
+         */
+        startBundle?: (fileName: string) => void;
+    }
+
+
     export interface BuildResults {
         buildMsg: string;
         totalTimeMs: number;
-        builtBundles: { fileName: string; timeMs: number; }[];
-        skippedBundles: { fileName: string; timeMs: number }[];
+        builtBundles: FileBundleResults[];
+        skippedBundles: FileBundleResults[];
+    }
+
+
+    export interface FileBundleResults {
+        fileName: string;
+        timeMs: number;
     }
 
 
@@ -57,16 +91,17 @@ module BrowserifyHelper {
      * The major reason to use this method instead of hand rolling the pipe() calls is the detailed error handling this method adds to each pipe() step.
      *
      * @param rebuildOnSrcChange flag indicating whether bundle should watch filesystem for changes and rebuild on change
+     * @param verbose whether to print bundle build info to 'gulp-util' log()
      * @param bundler the browserify object with a watchify plugin used to listener for 'update' events on to determine when to rebundle
      * @param getSourceStreams a function which creates the source stream and the one or more bundle output streams
      * @param additionalStreamPipes further transformations (i.e. [ (prevSrc) => prevSrc.pipe(vinyleSourceStream(...), (prevSrc) => prevSrc.pipe(gulp.dest(...)) ])
-     * @return a promise which completes when the first build completes and returns a message with the name of the compiled file and how long it took
+     * @param listeners various optional functions to call when bundle compile steps are completed, including a finishAll() function which passes back an object with stats about all the compiled bundles
      */
-    export function setupRebundleListener(rebuildOnSrcChange: boolean, bundler: browserify.BrowserifyObject,
+    export function setupRebundleListener(rebuildOnSrcChange: boolean, verbose: boolean, bundler: browserify.BrowserifyObject,
             getSourceStreams: (bundler: browserify.BrowserifyObject, updateEvent: any) => MultiBundleStreams,
-            additionalStreamPipes: [string, (prevStream: NodeJS.ReadableStream, streamOpts: BundleDst) => NodeJS.ReadableStream][]) {
-
-        var firstBuildDfd = Q.defer<BuildResults>();
+            additionalStreamPipes: [string, (prevStream: NodeJS.ReadableStream, streamOpts: BundleDst) => NodeJS.ReadableStream][],
+            listeners: BuildListeners) {
+        listeners = listeners || <BuildListeners>{};
 
         function rebundle(updateEvent?: any) {
             var expectedTotal = 0;
@@ -80,41 +115,63 @@ module BrowserifyHelper {
             function startCb(file: string) {
                 startTimes[file] = <number>Date.now();
                 expectDoneFiles.push(file);
-                gutil.log("start building '" + file + "'...");
+                if (verbose) {
+                    gutil.log("start building '" + file + "'...");
+                }
+                if (listeners.startBundle) {
+                    tryCall(listeners.startBundle, file);
+                }
             }
 
-            function doneCb(file: string, type: ("compile" | "skip")) {
+            function doneCb(srcName: string, file: string, type: ("compile" | "skip")) {
                 endTimes[file] = <number>Date.now();
-                if (type === "compile") { doneFiles.push(file); }
-                else if (type === "skip") { skippedFiles.push(file); }
+                if (type === "compile") {
+                    doneFiles.push(file);
+                    if (listeners.finishBundle) {
+                        tryCall(listeners.finishBundle, file);
+                    }
+                }
+                else if (type === "skip") {
+                    skippedFiles.push(file);
+                    if (listeners.skipBundle) {
+                        tryCall(listeners.skipBundle, file);
+                    }
+                }
                 else {
                     var errMsg = "invalid bundle completion type (expected: 'compile' or 'skip'): " + type;
-                    gutil.log(errMsg);
-                    firstBuildDfd.reject(errMsg);
+                    console.error(errMsg);
+                    if (listeners.error) {
+                        tryCall(listeners.error, srcName, file, errMsg);
+                    }
                     return;
                 }
 
                 var totalDone = doneFiles.length + skippedFiles.length;
                 if (totalDone >= expectedTotal) {
                     var endTime = <number>Date.now();
-                    var bldMsg = doneFiles.length > 0 ? "finished building: " + doneFiles.map((f) => f + " (" + (endTimes[file] - startTimes[file]) + " ms)").join(", ") : null;
-                    var skpMsg = skippedFiles.length > 0 ? "skipped building: " + skippedFiles.join(", ") : null;
-                    var buildMsg = "total time: " + (endTime - startTime) + " ms | " + (bldMsg ? bldMsg + (skpMsg ? " | " + skpMsg : "") : (skpMsg ? skpMsg : "no bundles"));
-                    gutil.log(buildMsg);
-                    firstBuildDfd.resolve({
-                        buildMsg,
-                        totalTimeMs: endTime - startTime,
-                        builtBundles: doneFiles.map((f) => ({ fileName: f, timeMs: (endTimes[f] - startTimes[f]) })),
-                        skippedBundles: skippedFiles.map((f) => ({ fileName: f, timeMs: (endTimes[f] - startTimes[f]) })),
-                    });
+                    var bldMsg = doneFiles.length > 0 ? doneFiles.map((f) => f + " (" + (endTimes[file] - startTimes[file]) + " ms)").join(", ") : null;
+                    var skpMsg = skippedFiles.length > 0 ? "skipped: " + skippedFiles.join(", ") : null;
+                    var buildMsg = "done building (" + (endTime - startTime) + " ms): " + (bldMsg ? bldMsg + (skpMsg ? " | " + skpMsg : "") : (skpMsg ? skpMsg : "no bundles"));
+                    if (verbose) {
+                        gutil.log(buildMsg);
+                    }
+                    if (listeners.finishAll) {
+                        tryCall(listeners.finishAll, {
+                            buildMsg,
+                            totalTimeMs: endTime - startTime,
+                            builtBundles: doneFiles.map((f) => ({ fileName: f, timeMs: (endTimes[f] - startTimes[f]) })),
+                            skippedBundles: skippedFiles.map((f) => ({ fileName: f, timeMs: (endTimes[f] - startTimes[f]) })),
+                        });
+                    }
                 }
             }
 
             function createErrorCb(srcName: string, dstFile: string) {
                 return function (err) {
-                    var msg = "error building '" + dstFile + "' at stream '" + srcName + "'";
-                    console.error(msg, err);
-                    firstBuildDfd.reject(msg + ": " + String(err));
+                    console.error("error building '" + dstFile + "' at stream '" + srcName + "'", err);
+                    if (listeners.error) {
+                        tryCall(listeners.error, srcName, dstFile, err);
+                    }
                 };
             }
 
@@ -125,7 +182,7 @@ module BrowserifyHelper {
                     var dstFilePath = bundle.dstFileName;
                     var resStream = bundle.stream;
                     if (resStream == null) {
-                        doneCb(dstFilePath, "skip");
+                        doneCb("initial-stream", dstFilePath, "skip");
                         return;
                     }
 
@@ -139,7 +196,7 @@ module BrowserifyHelper {
                         resStream.on("error", createErrorCb(streamName, dstFilePath));
                     }
 
-                    resStream.on("end", () => doneCb(dstFilePath, "compile"));
+                    resStream.on("end", () => doneCb(streamName, dstFilePath, "compile"));
 
                     startCb(dstFilePath);
                 });
@@ -152,14 +209,13 @@ module BrowserifyHelper {
             else {
                 startStreams(<BundleStream<NodeJS.ReadableStream>[]>bundles.bundleStreams);
             }
-
-            return firstBuildDfd.promise;
         }
 
         if (rebuildOnSrcChange) {
             bundler.on("update", rebundle);
         }
-        return rebundle();
+
+        rebundle();
     }
 
 
@@ -232,8 +288,27 @@ module BrowserifyHelper {
     }
 
 
+    function tryCall<T1, T2, T3>(func: (arg1: T1, arg2?: T2, arg3?: T3) => void, arg1: T1, arg2?: T2, arg3?: T3) {
+        if (func == null) { return; }
+        try {
+            switch(arguments.length - 1) {
+                case 3:
+                    return func(arg1, arg2, arg3);
+                case 2:
+                    return func(arg1, arg2);
+                case 1:
+                    return func(arg1);
+                default:
+                    throw new Error("unsupported number of arguments");
+            }
+        } catch (err) {
+            console.error("error calling '" + (func != null ? func.name : null) + "'", err);
+        }
+    }
+
+
     function isPromise<T>(p: any): p is Q.Promise<T> {
-        return Q.isPromiseAlike(p);
+        return p != null && typeof p.then === "function";
     }
 
 }