+++ bundlers/traceur/TraceurBundler.js
@@ -7,11 +7,15 @@ var Es6ifyToStream = require("./Es6ifyToStream");
  */
 var TraceurBundler;
 (function (TraceurBundler) {
-    function createTransformer(traceur, filePattern, traceurCompilerOpts, transformOpts) {
+    /** Create a browserify transform which compiles source files using traceur
+     */
+    function createTransformer(traceur, filePattern, traceurCompilerOpts, transformOpts, verbose) {
         Es6ifyToStream.traceurOverrides.global = true;
         // no file pattern, match all JS files
         var es6ifyCompile = Es6ifyToStream.createCompiler(traceur, filePattern, function (file, data) {
-            console.log("traceur: '" + PathUtil.toShortFileName(file) + "'"); // + ", data " + data.length + " done");
+            if (verbose) {
+                console.log("traceur: '" + PathUtil.toShortFileName(file) + "'"); // + ", data " + data.length + " done");
+            }
         });
         var res = {
             transform: function es6ifyTransform(file, opts) {